Replaced renderStyleguide SFC by <App /> and moved scrollTo logic into App
