+++ src/rsg-components/StyleGuide/StyleGuide.js
@@ -1,4 +1,3 @@
-import { isEqual, isFunction } from 'lodash';
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import TableOfContents from 'rsg-components/TableOfContents';
@@ -68,15 +67,6 @@ export default class StyleGuide extends Component {
 			slots: this.props.slots,
 			displayMode: this.props.displayMode,
 		};
-	}
-
-	componentDidUpdate(prevProps) {
-		// scroll to top of styleguide container when sections changed
-		if (this.props.pagePerSection && !isEqual(this.props.sections, prevProps.sections)) {
-			if (window && isFunction(window.scrollTo)) {
-				window.scrollTo(0, 0);
-			}
-		}
 	}
 
 	componentDidCatch(error, info) {
